requirements.txt
absl-py
tqdm
google-cloud-aiplatform
